pay-admin/pyproject.toml
[tool.poetry]
name = "pay-admin"
version = "0.1.0"
description = ""
authors = ["Travis Semple <travis8814@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
gunicorn = "^21.2.0"
flask = "^3.0.2"
flask-admin = "^1.6.1"
flask-oidc = "^2.1.1"
flask-sqlalchemy = "^3.1.1"
httplib2 = "^0.22.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
requests-futures = "^1.0.1"
wtforms = "^3.1.2"
werkzeug = "^3.0.1"
sqlalchemy = "^2.0.28"
itsdangerous = "^2.1.2"
jinja2 = "^3.1.3"
pay-api = {git = "https://github.com/bcgov/sbc-pay.git", subdirectory = "pay-api"}
pg8000 = "^1.30.5"
flask-session = {extras = ["filesystem"], version = "^0.8.0"}
pyjwt = "^2.8.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-mock = "^3.12.0"
requests = "^2.31.0"
pyhamcrest = "^2.1.0"
pytest-cov = "^4.1.0"
freezegun = "^1.4.0"
flake8 = "^7.0.0"
flake8-blind-except = "^0.2.1"
flake8-debugger = "^4.1.2"
flake8-docstrings = "^1.7.0"
flake8-isort = "^6.1.1"
flake8-quotes = "^3.4.0"
pep8-naming = "^0.13.3"
autopep8 = "^2.0.4"
coverage = "^7.4.3"
pylint = "^3.1.0"
pylint-flask = "^0.6"
pydocstyle = "^6.3.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
